+++ src/app/components/base-list.component.ts
@@ -12,6 +12,7 @@ import {ErrorHandlerService} from '@app/services/error-handler.service';
 import {LoadingOverlayService} from '@app/services/loading-overlay.service';
 import {LoggerService} from '@app/services/logger.service';
 import {UtilsService} from '@app/services/utils.service';
+import { compareNullableString } from '@app/utils/compare-nullable-string';
 import {config} from '@config';
 
 @Component({
@@ -176,7 +177,7 @@ export abstract class BaseListComponent<T extends Resource>
         );
         this.codelists.set(
           code,
-          [...list].sort((a, b) => a.description.localeCompare(b.description))
+          [...list].sort((a, b) => compareNullableString(a.description, b.description))
         );
       })
     );