+++ src/app/components/layers-permits/layers-permits-form/layers-permits-form.component.ts
@@ -28,6 +28,7 @@ import {ErrorHandlerService} from '@app/services/error-handler.service';
 import {LoadingOverlayService} from "@app/services/loading-overlay.service";
 import {LoggerService} from '@app/services/logger.service';
 import {UtilsService} from '@app/services/utils.service';
+import { compareNullableString } from '@app/utils/compare-nullable-string';
 import {constants} from '@environments/constants';
 
 
@@ -189,7 +190,7 @@ export class LayersPermitsFormComponent extends BaseFormComponent<CartographyGro
 
     this.permissionGroupTypes = this.codeList('cartographyPermission.type')
       .filter((item) => allowed.has(item.value))
-      .sort((a, b) => a.description.localeCompare(b.description));
+      .sort((a, b) => compareNullableString(a.description, b.description));
   }
 
   private defaultLayersPermitType(): CodeList | null {