+++ src/app/components/tasks-edit/task-form/task-edit-form.component.ts
@@ -47,6 +47,7 @@ import {ErrorHandlerService} from "@app/services/error-handler.service";
 import {LoadingOverlayService} from "@app/services/loading-overlay.service";
 import {LoggerService} from "@app/services/logger.service";
 import {UtilsService} from "@app/services/utils.service";
+import { compareNullableString } from '@app/utils/compare-nullable-string';
 import {magic} from "@environments/constants";
 
 /**
@@ -293,7 +294,7 @@ export class TaskEditFormComponent extends BaseFormComponent<TaskProjection> {
       throw new Error(`Task type ${this.taskTypeName} not found`);
     }
 
-    taskGroups.sort((a, b) => a.name.localeCompare(b.name));
+    taskGroups.sort((a, b) => compareNullableString(a.name, b.name));
     this.taskGroupList = taskGroups;
   }